# ras_orchestrator/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RAS-like Orchestrator API
  description: |
    API для взаимодействия с RAS-like оркестратором.
    Включает endpoints для приёма событий, управления политиками, мониторинга состояния.
  version: 0.1.0
  contact:
    name: RAS Team
    email: ras-team@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Локальный сервер разработки
  - url: https://api.ras.example.com
    description: Production сервер
...
paths:
...
  /policies/file/{filename}:
    get:
      tags:
        - policies
      summary: Получить содержимое файла политик
      description: Возвращает содержимое указанного YAML файла.
      operationId: getPolicyFile
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Имя файла (например, interrupt_policies.yaml)
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Файл не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /policies/evaluate:
    post:
      tags:
        - policies
      summary: Оценка политик для заданного контекста
      description: Оценивает политики определённого типа на основе предоставленного контекста.
      operationId: evaluatePolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateRequest'
      responses:
        '200':
          description: Результат оценки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateResponse'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Event ingestion failed."
        detail:
          type: string
          example: "Kafka broker unavailable"
        code:
          type: integer
          example: 500
    Policy:
      type: object
      properties:
        name:
          type: string
          example: "high_risk_security"
        version:
          type: string
          example: "1.0"
        enabled:
          type: boolean
          example: true
        priority:
          type: integer
          example: 10
        tags:
          type: array
          items:
            type: string
          example: ["security", "high_priority"]
        conditions:
          type: object
          additionalProperties: true
        actions:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
...
    EvaluateRequest:
      type: object
      required:
        - policy_type
        - context
      properties:
        policy_type:
          type: string
          example: "interrupt"
        context:
          type: object
          additionalProperties: true
          example:
            event:
              type: "security_alert"
              severity: "critical"
            salience:
              risk: 0.9
            current_mode: "normal"
            active_task_count: 3
    EvaluateResponse:
      type: object
      properties:
        matched:
          type: boolean
          example: true
        policies:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
        actions:
          type: object
          additionalProperties: true
          example:
            action: "interrupt"
            reason: "Высокий риск безопасности"
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API ключ для аутентификации